Respond to review comments.
1. Move Reorder into its own file with better Scaladoc comments.
2. Make Predicate's injections public, since predicate is public.
3. Improve some names and comments to be clearer.
4. A bit of other clean up.
