+++ brushfire-core/src/main/scala/com/stripe/brushfire/Brushfire.scala
@@ -57,7 +57,15 @@ case class Split[V, T](predicate: Predicate[V], leftDistribution: T, rightDistri
 
 /** Evaluates the goodness of a candidate split */
 trait Evaluator[V, T] {
-  /** returns a (possibly transformed) version of the input split, and a numeric goodness score */
+
+  /**
+   * Evaluate the fitness of a candidate split.
+   *
+   * This method may transform the split, in which case the score
+   * applies to the split that is returned. The result is optional to
+   * handle cases where a split is not valid (for example, if one side
+   * of a split is "empty" in some sense).
+   */
   def evaluate(split: Split[V, T]): Option[(Split[V, T], Double)]
 }
