+++ brushfire-core/src/main/scala/com/stripe/brushfire/Injections.scala
@@ -71,49 +71,50 @@ object JsonInjections {
     }
   }
 
+  implicit def predicateJsonInjection[V](implicit vInj: JsonNodeInjection[V], ord: Ordering[V] = null): JsonNodeInjection[Predicate[V]] = {
+    new AbstractJsonNodeInjection[Predicate[V]] {
+      def apply(pred: Predicate[V]) = {
+        val obj = JsonNodeFactory.instance.objectNode
+        pred match {
+          case IsPresent(None) => obj.put("exists", JsonNodeFactory.instance.nullNode)
+          case IsPresent(Some(pred)) => obj.put("exists", toJsonNode(pred)(predicateJsonInjection))
+          case EqualTo(v) => obj.put("eq", toJsonNode(v))
+          case LessThan(v) => obj.put("lt", toJsonNode(v))
+          case Not(pred) => obj.put("not", toJsonNode(pred)(predicateJsonInjection))
+          case AnyOf(preds) =>
+            val ary = JsonNodeFactory.instance.arrayNode
+            preds.foreach { pred => ary.add(toJsonNode(pred)(predicateJsonInjection)) }
+            obj.put("or", ary)
+        }
+        obj
+      }
+
+      override def invert(n: JsonNode) = {
+        n.getFieldNames.asScala.toList.headOption match {
+          case Some("eq") => fromJsonNode[V](n.get("eq")).map { EqualTo(_) }
+          case Some("lt") =>
+            if (ord == null)
+              sys.error("No Ordering[V] supplied but less than used")
+            else
+              fromJsonNode[V](n.get("lt")).map { LessThan(_) }
+          case Some("not") => fromJsonNode[Predicate[V]](n.get("not")).map { Not(_) }
+          case Some("or") => fromJsonNode[List[Predicate[V]]](n.get("or")).map { AnyOf(_) }
+          case Some("exists") =>
+            val predNode = n.get("exists")
+            if (predNode.isNull) Success(IsPresent[V](None))
+            else fromJsonNode[Predicate[V]](predNode).map(p => IsPresent(Some(p)))
+          case _ => sys.error("Not a predicate node: " + n)
+        }
+      }
+    }
+  }
+
   implicit def treeJsonInjection[K, V, T](
     implicit kInj: JsonNodeInjection[K],
     pInj: JsonNodeInjection[T],
     vInj: JsonNodeInjection[V],
     mon: Monoid[T],
     ord: Ordering[V] = null): JsonNodeInjection[Tree[K, V, T]] = {
-
-    implicit def predicateJsonNodeInjection: JsonNodeInjection[Predicate[V]] =
-      new AbstractJsonNodeInjection[Predicate[V]] {
-        def apply(pred: Predicate[V]) = {
-          val obj = JsonNodeFactory.instance.objectNode
-          pred match {
-            case IsPresent(None) => obj.put("exists", JsonNodeFactory.instance.nullNode)
-            case IsPresent(Some(pred)) => obj.put("exists", toJsonNode(pred)(predicateJsonNodeInjection))
-            case EqualTo(v) => obj.put("eq", toJsonNode(v))
-            case LessThan(v) => obj.put("lt", toJsonNode(v))
-            case Not(pred) => obj.put("not", toJsonNode(pred)(predicateJsonNodeInjection))
-            case AnyOf(preds) =>
-              val ary = JsonNodeFactory.instance.arrayNode
-              preds.foreach { pred => ary.add(toJsonNode(pred)(predicateJsonNodeInjection)) }
-              obj.put("or", ary)
-          }
-          obj
-        }
-
-        override def invert(n: JsonNode) = {
-          n.getFieldNames.asScala.toList.headOption match {
-            case Some("eq") => fromJsonNode[V](n.get("eq")).map { EqualTo(_) }
-            case Some("lt") =>
-              if (ord == null)
-                sys.error("No Ordering[V] supplied but less than used")
-              else
-                fromJsonNode[V](n.get("lt")).map { LessThan(_) }
-            case Some("not") => fromJsonNode[Predicate[V]](n.get("not")).map { Not(_) }
-            case Some("or") => fromJsonNode[List[Predicate[V]]](n.get("or")).map { AnyOf(_) }
-            case Some("exists") =>
-              val predNode = n.get("exists")
-              if (predNode.isNull) Success(IsPresent[V](None))
-              else fromJsonNode[Predicate[V]](predNode).map(p => IsPresent(Some(p)))
-            case _ => sys.error("Not a predicate node: " + n)
-          }
-        }
-      }
 
     implicit def nodeJsonNodeInjection: JsonNodeInjection[Node[K, V, T, Unit]] =
       new AbstractJsonNodeInjection[Node[K, V, T, Unit]] {
@@ -127,7 +128,7 @@ object JsonInjections {
           case SplitNode(k, p, lc, rc, _) =>
             val obj = JsonNodeFactory.instance.objectNode
             obj.put("key", toJsonNode(k))
-            obj.put("predicate", toJsonNode(p)(predicateJsonNodeInjection))
+            obj.put("predicate", toJsonNode(p)(predicateJsonInjection))
             obj.put("left", toJsonNode(lc)(nodeJsonNodeInjection))
             obj.put("right", toJsonNode(rc)(nodeJsonNodeInjection))
             obj
